line_crossing/requirements.txt
numpy==1.26.4
clearblade
python-dotenv
shapely==2.0.7
opencv-python==4.10.0.84